requirements.txt
Django==2.2.10
djangorestframework==3.12.4
python-dateutil==2.8.2
